+++ packages/swapper/src/swappers/cow/getCowSwapTradeQuote/getCowSwapTradeQuote.test.ts
@@ -1,13 +1,13 @@
-import { AssetService } from '@shapeshiftoss/asset-service'
-import { ethereum } from '@shapeshiftoss/chain-adapters'
+import { ethereum, FeeDataEstimate } from '@shapeshiftoss/chain-adapters'
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import { Asset } from '@shapeshiftoss/types'
+import { Asset, KnownChainIds } from '@shapeshiftoss/types'
 import Web3 from 'web3'
 
 import { GetEthTradeQuoteInput, TradeQuote } from '../../../api'
 import { ETH, FOX, WBTC, WETH } from '../../utils/test-data/assets'
 import { CowSwapperDeps } from '../CowSwapper'
 import { cowService } from '../utils/cowService'
+import { CowSwapQuoteApiInput } from '../utils/helpers/helpers'
 import { getCowSwapTradeQuote } from './getCowSwapTradeQuote'
 
 jest.mock('@shapeshiftoss/chain-adapters')
@@ -25,8 +25,26 @@ jest.mock('../utils/helpers/helpers', () => {
   }
 })
 
-const feeData = {
+const feeData: FeeDataEstimate<KnownChainIds.EthereumMainnet> = {
   fast: {
+    txFee: '4080654495000000',
+    chainSpecific: {
+      gasLimit: '100000',
+      gasPrice: '79036500000',
+      maxFeePerGas: '216214758112',
+      maxPriorityFeePerGas: '2982734547'
+    }
+  },
+  slow: {
+    txFee: '4080654495000000',
+    chainSpecific: {
+      gasLimit: '100000',
+      gasPrice: '79036500000',
+      maxFeePerGas: '216214758112',
+      maxPriorityFeePerGas: '2982734547'
+    }
+  },
+  average: {
     txFee: '4080654495000000',
     chainSpecific: {
       gasLimit: '100000',
@@ -37,7 +55,7 @@ const feeData = {
   }
 }
 
-const expectedApiInputWethToFox = {
+const expectedApiInputWethToFox: CowSwapQuoteApiInput = {
   appData: '0x0000000000000000000000000000000000000000000000000000000000000000',
   buyToken: '0xc770eefad204b5180df6a14ee197d99d808ee52d',
   from: '0x0000000000000000000000000000000000000000',
@@ -49,7 +67,7 @@ const expectedApiInputWethToFox = {
   validTo: 1656797787
 }
 
-const expectedApiInputWbtcToWeth = {
+const expectedApiInputWbtcToWeth: CowSwapQuoteApiInput = {
   appData: '0x0000000000000000000000000000000000000000000000000000000000000000',
   buyToken: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
   from: '0x0000000000000000000000000000000000000000',
@@ -105,11 +123,11 @@ const expectedTradeQuoteWbtcToWeth: TradeQuote<'eip155:1'> = {
   sellAssetAccountNumber: 0
 }
 
-const defaultDeps = {
+const defaultDeps: CowSwapperDeps = {
   apiUrl: '',
-  adapter: <ethereum.ChainAdapter>{},
-  web3: <Web3>{},
-  assetService: <AssetService>{}
+  adapter: {} as ethereum.ChainAdapter,
+  web3: {} as Web3,
+  feeAsset: WETH
 }
 
 describe('getCowTradeQuote', () => {
@@ -130,18 +148,16 @@ describe('getCowTradeQuote', () => {
   })
 
   it('should call cowService with correct parameters, handle the fees and return the correct trade quote when selling WETH', async () => {
-    const deps = {
+    const deps: CowSwapperDeps = {
       apiUrl: 'https://api.cow.fi/mainnet/api',
       adapter: {
         getAddress: jest.fn(() => Promise.resolve('address11')),
-        getFeeData: jest.fn(() => Promise.resolve(feeData))
+        getFeeData: jest.fn<Promise<FeeDataEstimate<KnownChainIds.EthereumMainnet>>, []>(() =>
+          Promise.resolve(feeData)
+        )
       } as unknown as ethereum.ChainAdapter,
-      web3: <Web3>{},
-      assetService: {
-        getAll: jest.fn(() => {
-          return { 'eip155:1/erc20:0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2': WETH }
-        })
-      } as unknown as AssetService
+      web3: {} as Web3,
+      feeAsset: WETH
     }
 
     const input: GetEthTradeQuoteInput = {
@@ -180,7 +196,7 @@ describe('getCowTradeQuote', () => {
   })
 
   it('should call cowService with correct parameters, handle the fees and return the correct trade quote when selling WBTC with undefined wallet', async () => {
-    const deps = {
+    const deps: CowSwapperDeps = {
       apiUrl: 'https://api.cow.fi/mainnet/api',
       adapter: {
         getAddress: jest.fn(() => {
@@ -188,12 +204,8 @@ describe('getCowTradeQuote', () => {
         }), // using this should throw an error
         getFeeData: jest.fn(() => Promise.resolve(feeData))
       } as unknown as ethereum.ChainAdapter,
-      web3: <Web3>{},
-      assetService: {
-        getAll: jest.fn(() => {
-          return { 'eip155:1/erc20:0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2': WETH }
-        })
-      } as unknown as AssetService
+      web3: {} as Web3,
+      feeAsset: WETH
     }
 
     const input: GetEthTradeQuoteInput = {