+++ packages/swapper/src/swappers/cow/getCowSwapTradeQuote/getCowSwapTradeQuote.ts
@@ -13,11 +13,14 @@ import {
   DEFAULT_ADDRESS,
   DEFAULT_APP_DATA,
   DEFAULT_SOURCE,
-  ORDER_KIND_SELL,
-  WETH_ASSET_ID
+  ORDER_KIND_SELL
 } from '../utils/constants'
 import { cowService } from '../utils/cowService'
-import { getNowPlusThirtyMinutesTimestamp, getUsdRate } from '../utils/helpers/helpers'
+import {
+  CowSwapQuoteApiInput,
+  getNowPlusThirtyMinutesTimestamp,
+  getUsdRate
+} from '../utils/helpers/helpers'
 
 export async function getCowSwapTradeQuote(
   deps: CowSwapperDeps,
@@ -25,7 +28,7 @@ export async function getCowSwapTradeQuote(
 ): Promise<TradeQuote<'eip155:1'>> {
   try {
     const { sellAsset, buyAsset, sellAmount, sellAssetAccountNumber, wallet } = input
-    const { adapter, assetService } = deps
+    const { adapter, feeAsset } = deps
 
     const { assetReference: sellAssetErc20Address, assetNamespace: sellAssetNamespace } =
       fromAssetId(sellAsset.assetId)
@@ -49,6 +52,18 @@ export async function getCowSwapTradeQuote(
       bnOrZero(sellAmount).lt(minQuoteSellAmount) ? minQuoteSellAmount : sellAmount
     )
 
+    const apiInput: CowSwapQuoteApiInput = {
+      sellToken: sellAssetErc20Address,
+      buyToken: buyAssetErc20Address,
+      receiver: DEFAULT_ADDRESS,
+      validTo: getNowPlusThirtyMinutesTimestamp(),
+      appData: DEFAULT_APP_DATA,
+      partiallyFillable: false,
+      from: DEFAULT_ADDRESS,
+      kind: ORDER_KIND_SELL,
+      sellAmountBeforeFee: normalizedSellAmount
+    }
+
     /**
      * /v1/quote
      * params: {
@@ -64,20 +79,11 @@ export async function getCowSwapTradeQuote(
      * }
      */
     const quoteResponse: AxiosResponse<CowSwapQuoteResponse> =
-      await cowService.post<CowSwapQuoteResponse>(`${deps.apiUrl}/v1/quote/`, {
-        sellToken: sellAssetErc20Address,
-        buyToken: buyAssetErc20Address,
-        receiver: DEFAULT_ADDRESS,
-        validTo: getNowPlusThirtyMinutesTimestamp(),
-        appData: DEFAULT_APP_DATA,
-        partiallyFillable: false,
-        from: DEFAULT_ADDRESS,
-        kind: ORDER_KIND_SELL,
-        sellAmountBeforeFee: normalizedSellAmount
-      })
+      await cowService.post<CowSwapQuoteResponse>(`${deps.apiUrl}/v1/quote/`, apiInput)
 
-    const { data } = quoteResponse
-    const quote = data.quote
+    const {
+      data: { quote }
+    } = quoteResponse
 
     const rate = bn(quote.buyAmount)
       .div(quote.sellAmount)
@@ -85,7 +91,6 @@ export async function getCowSwapTradeQuote(
       .toString()
 
     const receiveAddress = wallet ? await adapter.getAddress({ wallet }) : DEFAULT_ADDRESS
-    const wethAsset = assetService.getAll()[WETH_ASSET_ID]
 
     const [feeDataOptions, wethUsdRate, usdRateSellAsset] = await Promise.all([
       adapter.getFeeData({
@@ -93,7 +98,7 @@ export async function getCowSwapTradeQuote(
         value: '0',
         chainSpecific: { from: receiveAddress, contractAddress: sellAssetErc20Address }
       }),
-      getUsdRate(deps, wethAsset),
+      getUsdRate(deps, feeAsset),
       getUsdRate(deps, sellAsset)
     ])
 
@@ -102,7 +107,7 @@ export async function getCowSwapTradeQuote(
     const fee = bnOrZero(quote.feeAmount)
       .multipliedBy(bnOrZero(usdRateSellAsset))
       .div(bnOrZero(wethUsdRate))
-      .div(bn(10).exponentiatedBy(sellAsset.precision - wethAsset.precision))
+      .div(bn(10).exponentiatedBy(sellAsset.precision - feeAsset.precision))
       .toString()
 
     return {