+++ packages/swapper/src/swappers/utils/test-data/setupDeps.ts
@@ -1,7 +1,8 @@
-import { AssetService } from '@shapeshiftoss/asset-service'
 import { ethereum } from '@shapeshiftoss/chain-adapters'
 import * as unchained from '@shapeshiftoss/unchained-client'
 import Web3 from 'web3'
+
+import { WETH } from './assets'
 
 jest.mock('@shapeshiftoss/chain-adapters')
 jest.mock('web3')
@@ -23,7 +24,6 @@ export const setupDeps = () => {
 
   const ethNodeUrl = 'http://localhost:1000'
   const web3Provider = new Web3.providers.HttpProvider(ethNodeUrl)
-  const assetService = new AssetService()
 
-  return { web3: new Web3(web3Provider), adapter: ethChainAdapter, assetService }
+  return { web3: new Web3(web3Provider), adapter: ethChainAdapter, feeAsset: WETH }
 }